Moved CachedGLObject to glcommon
